fix(zero-cache): include the location hint when getting ReplicatorDO stub
